+++ CamadaBLL/CampanhaBLL.cs
@@ -127,7 +127,7 @@ namespace CamadaBLL
 				string query = db.CreateInsertSQL("tblCampanha");
 
 				//--- insert
-				return db.ExecutarInsertAndGetID(query);
+				return (int)db.ExecutarInsertAndGetID(query);
 			}
 			catch (Exception ex)
 			{
