+++ CamadaBLL/ReuniaoBLL.cs
@@ -147,7 +147,7 @@ namespace CamadaBLL
 				string query = db.CreateInsertSQL("tblReunioes");
 
 				//--- insert
-				return db.ExecutarInsertAndGetID(query);
+				return (int)db.ExecutarInsertAndGetID(query);
 
 			}
 			catch (Exception ex)
